Parsing Animations Resources
