+++ src/model/index.ts
@@ -1,7 +1,6 @@
 import * as THREE from 'three';
 
-import { getAnimIndex, Entity } from './entity';
-import { loadAnim } from './anim';
+import { Entity } from './entity';
 import {
     initSkeleton,
     createSkeleton,
@@ -39,7 +38,7 @@ export async function loadModel(params: any,
         getEntities(),
         getModels(bodyIdx, entityIdx),
         getModelsTexture(),
-        getAnimations(),
+        getAnimations(animIdx, entityIdx),
         loadLUTTexture()
     ]);
     const resources = { ress, pal, entities, body, texture, anim };
@@ -77,6 +76,7 @@ function loadModelData(params: any,
     const entities = resources.entities;
     const texture = resources.texture;
     const body = resources.body;
+    const anim = resources.anim;
 
     const model = {
         palette,
@@ -93,9 +93,6 @@ function loadModelData(params: any,
         entity: entities[entityIdx],
         materials: []
     };
-
-    const realAnimIdx = getAnimIndex(model.entity, animIdx);
-    const anim = loadAnim(model, model.anims, realAnimIdx);
 
     const skeleton = createSkeleton(body);
     initSkeleton(animState, skeleton, anim.loopFrame);
