+++ src/ui/editor/areas/model/ModelEditorContent.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as THREE from 'three';
 import { each } from 'lodash';
+
 import Renderer from '../../../../renderer';
 import { fullscreen } from '../../../styles/index';
 import FrameListener from '../../../utils/FrameListener';
@@ -11,7 +12,6 @@ import {
     updateKeyframeInterpolation
 } from '../../../../model/animState';
 import { getAnim } from '../../../../model/entity';
-import { loadAnim } from '../../../../model/anim';
 import DebugData from '../../DebugData';
 import fmod from './utils/fmod';
 import {get3DOrbitCamera} from './utils/orbitCamera';
@@ -19,6 +19,7 @@ import { TickerProps } from '../../../utils/Ticker';
 import {
     registerResources,
     preloadResources,
+    getAnimations,
 } from '../../../../resources';
 import { loadEntities } from './browser/entitities';
 
@@ -209,7 +210,7 @@ export default class Model extends FrameListener<Props, State> {
         this.zoom = Math.min(Math.max(-1, this.zoom), 8);
     }
 
-    frame() {
+    async frame() {
         const { renderer, animState, clock, model, scene, grid } = this.state;
         const { entity, body, anim, rotateView, wireframe } = this.props.sharedState;
         if (this.entity !== entity || this.body !== body) {
@@ -236,7 +237,7 @@ export default class Model extends FrameListener<Props, State> {
         };
         renderer.stats.begin();
         if (model) {
-            const interpolate = this.updateModel(
+            const interpolate = await this.updateModel(
                 model,
                 animState,
                 entity,
@@ -250,13 +251,13 @@ export default class Model extends FrameListener<Props, State> {
         renderer.stats.end();
     }
 
-    updateModel(model, animState, entityIdx, animIdx, time) {
+    async updateModel(model, animState, entityIdx, animIdx, time) {
         const entity = model.entities[entityIdx];
         const entityAnim = getAnim(entity, animIdx);
         let interpolate = false;
         if (entityAnim !== null) {
             const realAnimIdx = entityAnim.animIndex;
-            const anim = loadAnim(model, model.anims, realAnimIdx);
+            const anim = await getAnimations(animIdx, entityIdx);
             animState.loopFrame = anim.loopFrame;
             if (animState.prevRealAnimIdx !== -1 && realAnimIdx !== animState.prevRealAnimIdx) {
                 updateKeyframeInterpolation(anim, animState, time, realAnimIdx);
